cleanup consistent hash cmd
